backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
python-dotenv
better-auth
openai>=1.3.0
mcp
